Update RegEx to support longer URLs
The bugfix I introduced in https://github.com/github/paste-markdown/pull/65
caused a regression because we were no longer accounting for paths after the
initial slash at the end of the domain. This updates the RegEx to account for
those paths.

I tried to use `new URL(potentialUrl)` to do the validation, but that would still match
when content was after the URL, because the spaces were getting escaped, so I stuck with the
RegExp (and yes, now I have two problems).